requirements.txt
Flask>=3.0
gunicorn
Werkzeug>=3.0
itsdangerous>=2.1
click>=8.1
Jinja2>=3.1
blinker>=1.6
psycopg2-binary